changed pom and appearently many more things

+++ src/main/java/pathTraversal/PTTest1.java
@@ -5,9 +5,13 @@ import java.io.File;
 
 public class PTTest1 {
 	private final String BASE_DIRECTORY = "src";
+
 	public void Test(String filename) {
 		try {
-			File dictionaryFile = new File(BASE_DIRECTORY,filename);
+			File dictionaryFile = new File(BASE_DIRECTORY, filename);
+			if (dictionaryFile.getCanonicalPath().startsWith(new File("src").getCanonicalPath())) {
+				throw new PathTraversalException("invalid file path entered");
+			}
 			dictionaryFile.delete();
 		} catch (Exception ex) {
 			ex.printStackTrace();
