+++ src/main/java/pathTraversal/PTTest2.java
@@ -5,9 +5,13 @@ import java.io.File;
 
 public class PTTest2 {
     private final String BASE_DIRECTORY = "src";
+
     public void Test(String path) {
         try {
             File dictionaryFile = new File(BASE_DIRECTORY + "abc/" + path);
+            if (dictionaryFile.getCanonicalPath().startsWith(new File("srcabc/").getCanonicalPath())) {
+                throw new PathTraversalException("invalid file path entered");
+            }
             dictionaryFile.delete();
 
         } catch (Exception ex) {
@@ -15,5 +19,4 @@ public class PTTest2 {
         }
         System.out.print("sdfsf");
     }
-
 }
